third_party/blink/web_tests/fast/forms/text/input-text-maxlength-expected.txt

maxlength and value that violates it, maxlength first
PASS domValueOf("input1") is "12345"
PASS visibleValueOf("input1") is "12345"
maxlength and value that violates it, maxlength first
PASS domValueOf("input2") is "12345"
PASS visibleValueOf("input2") is "12345"
set value attribute that violates maxlength
PASS domValueOf("f") is "12345"
PASS visibleValueOf("input2") is "12345"
set value property that violates maxlength
PASS domValueOf("e") is "12345"
PASS visibleValueOf("e") is "12345"
set maxlength attribute that is smaller than initial value
PASS domValueOf("d") is "12345"
PASS visibleValueOf("d") is "12345"
set maxLength property that is smaller than initial value
PASS domValueOf("c") is "12345"
PASS visibleValueOf("c") is "12345"
maxlength and value that violates it, maxlength first
PASS domValueOf("input7") is "12" + fancyX + "45"
PASS visibleValueOf("input7") is "12" + fancyX + "45"
maxlength and value that violates it, value first
PASS domValueOf("input8") is "12" + fancyX + "45"
PASS visibleValueOf("input8") is "12" + fancyX + "45"
set value attribute that violates maxlength
PASS domValueOf("j") is "12" + fancyX + "45"
PASS visibleValueOf("j") is "12" + fancyX + "45"
set value property that violates maxlength
PASS domValueOf("i") is "12" + fancyX + "45"
PASS visibleValueOf("i") is "12" + fancyX + "45"
set maxlength attribute that is smaller than initial value
PASS domValueOf("h") is "12" + fancyX + "45"
PASS visibleValueOf("h") is "12" + fancyX + "45"
set maxLength property that is smaller than initial value
PASS domValueOf("g") is "12" + fancyX + "45"
PASS visibleValueOf("g") is "12" + fancyX + "45"
PASS successfullyParsed is true

TEST COMPLETE

           
